2019.11.04-22:30 : Preliminary AGA Dual Playfield Support
- Updated +W.s/EcCopHo method by inserting BplCon3 register to allow 2nd Dual Playfield Screen to use colors 16-31 [Temporar change]
- execute aData to update Editor_Config
- execute aw to update Amos.library
- Updated the new DualPlayfield scrolling test demo.

Preliminary support for Dual Playfield with 2 x 16 colors screen is available.
Don't forget that it's a WIP/Beta version. Don't use it to develop products as it may be unstable.

PS : Make a copy of your existing LIBS:AMOS.Library somewhere else to have no AMOS.Library in LIBS: directory as the WIP one is in the APSystem directory.
Don't overwrite your existing AMOS Professional configuration with this one, make a copy of your own before testing this one.